Add files via upload
The last piece of this is a modification of the echonestExample.py - as long as we input a dataframe with billboard info, for either just a week, or multi-year span, this should gather up all the available info and store it in a csv file. We can scale this up later.